requirements.txt
altgraph==0.17
asgiref==3.2.10
auto-py-to-exe==2.7.5
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.40.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
Eel==0.12.4
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
install==1.3.3
itypes==1.2.0
Jinja2==2.11.2
macholib==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyInstaller==3.6
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.1.2
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whichcraft==0.6.1
whitenoise==5.1.0
zope.event==4.4
zope.interface==5.1.0
